Bug 43201: explicit order of control elements
decided on the elements to include based on the order in the array but
insert them into the dom based on the value of position. this assures
that the menu is always on the right but volume controls are showing up
for smaller players.

Change-Set 2: Added some comments for clarity that position only sets
order, not inclusion.
* Use base 10 for pos, so that its clear its not
a pixel value and extensions can always add float decimal points.
* Adds backward compatibility with player modules that don't specify
position.

Change-Id: Ieeceead216d526b450d8ce8786d5a57251e4cc6d