Clean up RL module definitions.
* Remove unneeded addModuleStyles(), it's redundant with addModule().
* Replace two array dependencies with their single contents.
* Use __DIR__ instead of dirname.
* Use + for concatenate $baseExtensionResource instead of array_merge().
* Improve a hook comment.

Related to bug 55550 though this doesn't reduce amount of JS code loaded.

I couldn't get TMH to work on my wiki, so this is untested, but the modules
load without errors in 1.22.
TimedMediaHandler claims to work with MediaWiki 1.17, so some of these may
be intentional for backwards compatibility and the cleanup unwanted.

Change-Id: Ia7b465c2db154620585db481db88a4c88a002038