Make sure jquery.ui.dialog is loaded if there are pop up videos
Follow up d20058ed77. If jquery.ui.dialog isn't loaded at
the beginning, then when someone clicks on a pop up video, the
pop up dialog has the wrong dimensions. This issue has not
appeared on Wikimedia, as other scripts are loading jquery.ui.dialog.

Not entirely sure if I took the right approach for loading it, but
this seemed like the most straightforward way to do it and have it
only on pages where needed.

Bug: 72437
Change-Id: Ia903cded423a8505528425e15abb1c87513f1c49