Remove z-index on cannot play file warning
This was causing the text in the warning to be unselectable, and
unclickable. Given that commons has links in that message, this
is a pretty severe usability issue.

I don't know why the z-index was set on this element. If it was
important, giving it a higher z-index (It was 2, would need to
be > 1002 to be above the containing div) would also fix the
issue.

Change-Id: Ibd95d533b91f1f8b6bb9c983486e849c95005e8b