Fix ?embedplayer=yes mode, which is currently totally broken
This is a little bit of a voodoo fix, as I don't entirely
understand the js of TimedMediaHandler. This also fixes some
deprecation warnings. I'm submitting this despite the vodooness
of the fix, since the behaviour on production is really broken.

Seems this was broken by fa5ea5be2370, but may also have had
additional causes.

Bug: 66143
Change-Id: Ie777a2f75a8a661569ee3dd25811621c9b459b80