Only do pop up video player if we actually have a bigger size
We have a pop up dialog box video player for cases where the
user embedded a small thumbnail. However it can still go
off in cases where the thumbnail is as big as the source video,
which doesn't make a lot of sense.

Also changes the comparision to the minimun video size to be
< instead of <=, since that's more consistent with how its
documented. I believe that's also more consistent with how its
used. For example, Wikimedia has it set to 800px, which is same
size as image page display. Presumably the intention was for the
dialog to not trigger on the image description page.

Change-Id: I25b213c28a0028c37b1ad5dfd43dd9d8f0ce3445