Handle double tags properly (fix php warning about trim())
A rarely used feature of ogg comments is that you can specify
a comment has multiple values, by including the comment twice.
The library used to get metadata handled that situation correctly,
but the code that outputted the metadata assumed that the metadata
comments were always a string and not an array.

Fixes Warning: trim() expects parameter 1 to be string...

Bug: 66448
Change-Id: I1325a246c1152b91ec011d1ff13db66288c9b22b