Work in progress: ogv.js media player for Safari/IE (2 of 3)
Plugin to MwEmbedPlayer for using ogv.js JavaScript playback
of Ogg Theora/Vorbis media in Safari and IE.

The actual JS libraries are in the previous commit in order
to make this commit easier to review.

This commit adds only the plugin for the desktop player and
support code that's common with the mobile player coming in
the next commit.

Bug: T63823
Change-Id: Ia6b2e66864a596dbb4ba5841b9da135021f38276