Work in progress: ogv.js media player for Safari/IE (3 of 3)
Mobile video/audio-player overlay, supporting both native and
ogv.js JavaScript-based playback (used for Safari).

There is a bug that when using ogv.js playback, the playback
doesn't stop when closing the overlay. There is a commented-out
'hide' method on the overlay which fixes the stop, but breaks
subsequent uses of the overlay. Not sure why...

Note ogv.js doesn't work in IE for Windows Phone, which currently
lacks both Web Audio and Flash support.

Bug: 61823
Change-Id: Ib0083dc6b50eb91adb9e9624a5e90448860441e2