WIP - optional MJPEG/mu-law MOV/WAV transcodes for iOS fallback
iOS supports native playback of MOV and WAV files with a VERY limited
set of non-patent-encumbered codecs: Motion-JPEG video and mu-law PCM
audio.

MJPEG is very inefficient because every frame is a reference frame,
so it's necessary to scale back the resolution, frame rate, and
quality in order to fit into roughly the same bandwidth as the
existing 360p WebM and Theora transcodes.

Audio is reduced to 22.5 KHz monophonic with mu-law encoding.

(Since these do play natively, they should work in the mobile view,
with JavaScript disabled, or in screen-scraped content in iOS apps.
Better picture and sound quality can be obtained using ogv.js instead,
but this is CPU-heavy on iOS and needs more plumbing to be hooked up.)

Activate like so:

  $wgEnabledTranscodeSet[] = WebVideoTranscode::ENC_MJPEG_160P;
  $wgEnabledAudioTranscodeSet[] = WebVideoTranscode::ENC_WAV_ULAW;

Todo:
* fix playback in desktop mode

Change-Id: I4f78211c15f3bad2663a5f21be4f4c3210760b3f
Task: T101716