WIP - optional MJPEG/mu-law MOV/WAV transcodes for iOS fallback
iOS supports native playback of MOV and WAV files with a VERY limited
set of codecs: Motion-JPEG video and mu-law PCM audio. MJPEG is very
inefficient because every frame is a reference frame, so quality
is quite bad in order to fit into the same bandwidth as the existing
360p WebM and Theora transcodes.

Activate like so:

  $wgEnabledTranscodeSet[] = WebVideoTranscode::ENC_MJPEG_160P;
  $wgEnabledTranscodeSet[] = WebVideoTranscode::ENC_MJPEG_240P;
  $wgEnabledTranscodeSet[] = WebVideoTranscode::ENC_MJPEG_360P;
  $wgEnabledAudioTranscodeSet[] = WebVideoTranscode::ENC_WAV_ULAW;

Currently there are three resolution options, all with low output
quality due to trying to fit into the same bandwidth as the 360p
Theora and WebM transcodes. Intend to test a bit and then settle
on one setting.

Note that iOS seems to refuse to play back files using a video
quality setting; have to just set the bandwidth max and go from there.

Note that iOS plays AVIs with MJPEG as well, but does not stream
them in the browser. QuickTime MOV with MJPEG does stream, which
is better for the poor user.

Todo:
* settle on a resolution for MJPEG
* fix playback in desktop mode

Change-Id: I4f78211c15f3bad2663a5f21be4f4c3210760b3f
Task: T101716