ogv.js media player for desktop Safari/IE/Edge (2 of 2)
Plugin to MwEmbedPlayer for using ogv.js JavaScript playback
of Ogg Theora/Vorbis media in Safari, Edge and IE.

The actual JS libraries are in a previous commit in order
to make this commit easier to review.

Note that while the player works in Safari for iOS, this commit
does not cover MobileFrontend integration for TimedMediaHandler,
which will come separately (still under development).

Updated to ogv.js 0.9.3 with basic threading for the decods;
performs smoother on slow machines, but may introduce new bugs.

Bug: T63823
Change-Id: Ia6b2e66864a596dbb4ba5841b9da135021f38276