Add support for WebM VP9/Opus transcodes
Adds transcode definitions for VP9/Opus transcodes at half the
video bitrate of VP8/Vorbis. HD transcodes use tiled encoding
mode which will help some decoders to parallelize playback.

Not yet enabled by default, so this should be safe to merge.

Requires ffmpeg or avconv with vp9 and opus support compiled in;
does not yet work with the ffmpeg in MediaWiki-Vagrant or in
Wikimedia production!

To test, set up suitable ffmpeg & enable the transcodes manually:

$wgEnabledTranscodeSet[] = WebVideoTranscode::ENC_VP9_360P;
$wgEnabledTranscodeSet[] = WebVideoTranscode::ENC_VP9_480P;
$wgEnabledTranscodeSet[] = WebVideoTranscode::ENC_VP9_720P;
$wgEnabledTranscodeSet[] = WebVideoTranscode::ENC_VP9_1080P;

VP9 will take priority if transcodes are available and the browser
supports playback; works in Firefox 39 and Chrome 44.

Bug: T63805
Change-Id: Idbea47fbde2a8862084a7e77eb1ca928083d5b85