Fix tests in TimedMediaHandler
This adds two missing tests to the required list in TimedMediaHandler, It
adds ApiTestCaseVideoUpload.php and TestWebMHandler.php.

Also fix class in TestWebMHandler.php which had class TestOggHandler which
was used by another file in TimedMediaHandler so it would fail the jenkins
tests so changed the class to TestWebMHandler which now passes jenkins.

Also update the test to support some recent changes done to
TimedMediaHandler.

Change-Id: Ibe4a423e0bcb1f2a416707609fa42b41d8f90b83