Expand trigger area for play button contrast change to whole player
The play button is partially transparent and can be hard to notice
on certain backgrounds. It becomes brighter when moused over which
is not much help in finding it. Make it change color when the whole
player is moused over instead (this is how e.g. YouTube handles
button color).

This also makes color change acually work for thumbnails
(.PopUpMediaTransform) in Chrome, before it was always the non-hover
style that was applied. I could not figure out the reason for that,
seemed like a Chrome bug.

Bug: T117344
Change-Id: I3f141d8854eaa9f4a4d87ec80555a25fc81a2ec7