Automatically re-add transcode jobs if transcode pending for 72h
I'm not 100% sure this is actually a good idea, but here goes.
There's currently a large number of pending transcodes at commons,
which show a status of waiting for job to run. However it claims
to be waiting for days despite the fact that the video scalars
appear to be idle.

Thus this checks if the transcode is pending for more than 72
hours. If so, it checks if job queue empty, and if it is,
re-inserts the job.

Note, this would potentially cause isEmpty() to be called a lot,
as I believe the transcode insertion code runs on page view of
image page.

Bug: T63401
Bug: T49312
Change-Id: Ie11b4b57013a7be5d51114f9781825e5e23a03cc