[Untested] Trigger transcode jobs from FileUpload hook
We only do transcodes for local files. The previously used
UploadComplete hook is really just a wrapper around this FileUpload
hook, but then not specific to LocalFile. Using this lower level hook,
allows it to be triggered by importImages.php, which only uploads
local files using the lower level recordUpload2.

Bug: T127597
Change-Id: Id79fad222d8d27ce7ecf77b7eb59bca54bebff3c