Allow audio player to be larger than 220px
This is a stupid mistake that normalized the audio player to use a
default of 220px as a 'natural' width and then restricted the
thumbnailing to be no larger than the 'natural' size. As such audio
players were never allowed to be larger than 220px.
Audio players under 220px in many conditions will not have a seekbar,
since there is not enough room on the controller to present this
seekbar.

Bug: T56370
Change-Id: I6362076e8a24234d2350bc6eb54c81fa9fef04cc