Shut down database connections during transcode shell-outs
Transcoding via ffmpeg/ffmpeg2theora can be quite slow, especially
on very large video files it may well take many minutes or even a
couple hours.

Shut down the database connections in the transcode jobs to avoid
leaving long-running connections or transactions which can cause
trouble.

This is done by calling LBFactory::shutdownInstance(), which closes
out all connections to all databases. When we need to load another
database connection for further work after the transcode is done,
we just call wfGetDB() and it dives down to the LBFactory behind
the scenes to reconnect.

Bug: T97641
Change-Id: I935ca8ec465542a12016924dff6f73eb91df8f8d