Don't specify controls, because we show JS controls
Learned something new from the spec. We don't need this attribute in
the DOM, not even for situations where user-agent does not have JS.

"If the [controls] attribute is present, or if scripting is disabled
for the media element, then the user agent should expose a user
interface to the user. This user interface should include features to
begin playback, pause playback, seek to an arbitrary position in the
content (if the content supports arbitrary seeking), change the
volume, change the display of closed captions or embedded
sign-language tracks, select different audio tracks or turn on audio
descriptions, and show the media content in manners more suitable to
the user (e.g. fullscreen video or in an independent resizable
window). Other controls may also be made available."

Change-Id: I52f339c5e000928e224c507312e45aade549b25f