Apply ext.tmh.player.styles w/ videojs styles on mobile
Was missing the target specifier on the ext.tmh.player.styles.

Had to add a workaround for Minerva skin for thumbnailed videos,
as Minerva stuff rewrites some of the surrounding div styles to
use 'width: auto', which apparently doesn't handle properly the
case where the .video-js block is 'display: inline-block'. In
this case it's safe to use 'display: block' since it's in a block
anyway, so just added an override back. :)

Change-Id: Ifd50dd4aab29997fc1345854809997460ef280d3