Fix fatal for TimedMediaHandler previews without files.
The TimedMediaHandler extension is used to create playback thumbnails
for ogg files rendered by the Score extension. In this case, the
TimedMediaHandler has no file connected to it, so code inside the
thumbnail renderer and any code called from that renderer, should not
depend on the presence of a source file.

Follow-up to: I4227b88
Bug: T135491

Change-Id: I62b02a73679b82b0feae92fa244b57e782c51385