Don't preload entire video files in video.js mode
Uses preload="metadata" instead of preload="auto" for native
video/audio playback, or preload="none" for ogv.js playback.

Prevents aggressive preloading of video and audio files that
might not get used. Especially with ogv.js backend this can
fire off a lot of extra requests and threads.

Note that with preload="metadata" there will be fetches for
the file on desktop (but not on mobile, which ignores the
preload hint) but only the very beginning of the file will
be read in.

Change-Id: Ib71fd0397df92e245b9deda41acf57b137e0000b